move mocks dir to root
