+++ mocks/DuplexStream.ts
