+++ src/MetaMaskInpageProvider.rpc.test.ts
@@ -1,6 +1,6 @@
+import MockDuplexStream from '../mocks/DuplexStream';
 import MetaMaskInpageProvider from './MetaMaskInpageProvider';
 import messages from './messages';
-import MockDuplexStream from './mocks/DuplexStream';
 
 const MOCK_ERROR_MESSAGE = 'Did you specify a mock return value?';
