+++ tsconfig.json
@@ -12,5 +12,5 @@
     "typeRoots": ["./node_modules/@types"]
   },
   "include": ["./src/*.ts"],
-  "exclude": ["**/*.test.ts"]
+  "exclude": ["**/*.test.ts", "mocks/"]
 }
